Windows and tray are now handled serverside with the WindowsHandler class
